requirement.txt
streamlit
opencv-python
pillow
torch
torchvision
transformers
faiss-cpu
numpy